# com.primeton/generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "我的测试",
    "version" : "v1",
    "title" : "宠物信息"
  },
  "host" : "127.0.0.1:8080",
  "tags" : [ {
    "name" : "宠物信息相关api"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/pet" : {
      "post" : {
        "tags" : [ "宠物信息相关api" ],
        "summary" : "添加宠物信息",
        "description" : "",
        "operationId" : "insertPet",
        "produces" : [ "text/html" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PetDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          },
          "400" : {
            "description" : "请求参数没填好"
          },
          "404" : {
            "description" : "请求路径没有或页面跳转路径不对"
          }
        }
      },
      "put" : {
        "tags" : [ "宠物信息相关api" ],
        "summary" : "更新宠物信息",
        "description" : "",
        "operationId" : "updatePet",
        "produces" : [ "text/html" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PetDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          },
          "400" : {
            "description" : "请求参数没填好"
          },
          "404" : {
            "description" : "请求路径没有或页面跳转路径不对"
          }
        }
      }
    },
    "/pet/all" : {
      "get" : {
        "tags" : [ "宠物信息相关api" ],
        "summary" : "获取所有宠物信息",
        "description" : "",
        "operationId" : "getPetAll",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PetDto"
              }
            }
          },
          "404" : {
            "description" : "请求路径没有或页面跳转路径不对"
          }
        }
      }
    },
    "/pet/{id}" : {
      "get" : {
        "tags" : [ "宠物信息相关api" ],
        "summary" : "通过ID获取宠物信息",
        "description" : "",
        "operationId" : "getPet",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PetDto"
            }
          },
          "400" : {
            "description" : "请求参数没填好"
          },
          "404" : {
            "description" : "请求路径没有或页面跳转路径不对"
          }
        }
      },
      "delete" : {
        "tags" : [ "宠物信息相关api" ],
        "summary" : "通过ID删除宠物信息",
        "description" : "",
        "operationId" : "deletePet",
        "produces" : [ "text/html" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          },
          "400" : {
            "description" : "请求参数没填好"
          },
          "404" : {
            "description" : "请求路径没有或页面跳转路径不对"
          }
        }
      }
    }
  },
  "definitions" : {
    "Category" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
    "PetDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "category" : {
          "$ref" : "#/definitions/Category"
        },
        "name" : {
          "type" : "string"
        },
        "tag" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Tag"
          }
        },
        "status" : {
          "type" : "string"
        }
      }
    },
    "Tag" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string"
        }
      }
    }
  }
}